Refactor model reference hooks and imports
- Rename hook files to use kebab-case naming convention
- Remove debug console logs from model reference data API
- Update import paths in components to match new hook file names
- Remove unused hooks: useAssetClassFilter, useModelReferenceData, useUseFilter
