+++ frontend/src/features/models/components/model-reference-form.tsx
@@ -5,9 +5,9 @@ import { zodResolver } from '@hookform/resolvers/zod'
 import { useForm, useFieldArray } from 'react-hook-form'
 import * as z from 'zod'
 import { PlusCircle, Trash2, Loader2 } from 'lucide-react'
-import { useSeparateModelReferenceData } from '../hooks/useModelReferenceData'
-import { useAssetClassFilter } from '../hooks/useAssetClassFilter'
-import { useUseFilter } from '../hooks/useUseFilter'
+import { useSeparateModelReferenceData } from '../hooks/use-model-reference-data'
+import { useAssetClassFilter } from '../hooks/use-asset-class-filter'
+import { useUseFilter } from '../hooks/use-use-filter'
 
 import { Button } from '@/components/ui/button'
 import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form'
