+++ instructions/featured-pattern.md
@@ -1,0 +1,171 @@
+# Feature-Based Architecture Pattern
+
+## Overview
+
+The feature-based architecture pattern organizes code by business features rather than technical layers. This approach creates self-contained modules that encapsulate all aspects of a specific feature, making the codebase more maintainable, scalable, and easier to understand.
+
+## Directory Structure
+
+A typical feature module follows this structure:
+
+```
+src/
+└── features/
+    └── feature-name/
+        ├── api/
+        │   ├── create-resource.ts
+        │   ├── delete-resource.ts
+        │   ├── get-resource.ts
+        │   ├── get-resources.ts
+        │   └── update-resource.ts
+        ├── components/
+        │   ├── create-resource.tsx
+        │   ├── delete-resource.tsx
+        │   ├── resource-view.tsx
+        │   ├── resources-list.tsx
+        │   └── update-resource.tsx
+        └── types/
+            └── index.ts
+```
+
+## Key Components
+
+### API Layer
+
+The API layer handles all data fetching and mutations related to the feature. Each file typically exports:
+
+1. A base function that performs the API call
+2. A query/mutation options function (for React Query)
+3. A hook that uses the options function
+
+Example from `get-resources.ts`:
+
+```typescript
+// Base function
+export const getResources = (params): Promise<Response> => {
+	return api.get(`/resources`, { params })
+}
+
+// Query options
+export const getResourcesQueryOptions = params => {
+	return queryOptions({
+		queryKey: ['resources', params],
+		queryFn: () => getResources(params)
+	})
+}
+
+// Hook
+export const useResources = options => {
+	return useQuery({
+		...getResourcesQueryOptions(options.params),
+		...options.queryConfig
+	})
+}
+```
+
+### Components Layer
+
+The components layer contains all UI components specific to the feature. These components:
+
+1. Use the API hooks to fetch and mutate data
+2. Handle UI state and user interactions
+3. Implement feature-specific business logic
+
+Components are typically organized by their purpose:
+
+- List views (e.g., `resources-list.tsx`)
+- Detail views (e.g., `resource-view.tsx`)
+- Create/update forms (e.g., `create-resource.tsx`, `update-resource.tsx`)
+- Delete confirmations (e.g., `delete-resource.tsx`)
+
+### Types Layer
+
+The types layer defines TypeScript interfaces and types specific to the feature. This includes:
+
+1. API response types
+2. Form input types
+3. Component prop types
+4. Feature-specific domain types
+
+## Implementation Examples
+
+### Tables Feature Example
+
+Our `tables` feature follows this pattern:
+
+```
+features/
+└── tables/
+    ├── api/
+    │   ├── create-table-row.ts
+    │   ├── delete-table-row.ts
+    │   ├── get-table-data.ts
+    │   ├── get-tables.ts
+    │   └── update-table-row.ts
+    ├── components/
+    │   └── table-editor/
+    │       ├── components/
+    │       │   ├── table-sidebar.tsx
+    │       │   ├── table-view.tsx
+    │       │   ├── table-toolbar.tsx
+    │       │   └── table-pagination.tsx
+    │       ├── index.tsx
+    │       └── types.ts
+    └── types/
+        └── index.ts
+```
+
+### Discussions Feature Example
+
+The `discussions` feature demonstrates the pattern with:
+
+```
+features/
+└── discussions/
+    ├── api/
+    │   ├── create-discussion.ts
+    │   ├── delete-discussion.ts
+    │   ├── get-discussion.ts
+    │   ├── get-discussions.ts
+    │   └── update-discussion.ts
+    └── components/
+        ├── create-discussion.tsx
+        ├── delete-discussion.tsx
+        ├── discussion-view.tsx
+        ├── discussions-list.tsx
+        └── update-discussion.tsx
+```
+
+## Best Practices
+
+1. **Encapsulation**: Keep feature-specific code within the feature directory.
+2. **Consistency**: Follow the same pattern across all features.
+3. **Reusability**: Extract common patterns to shared utilities when appropriate.
+4. **Separation of Concerns**:
+   - API layer handles data fetching and mutations
+   - Components layer handles UI and user interactions
+   - Types layer defines the feature's data structures
+5. **Naming Conventions**: Use consistent, descriptive names for files and functions.
+6. **Minimal Dependencies**: Minimize dependencies between features.
+
+## Benefits
+
+1. **Modularity**: Features are self-contained and can be developed independently.
+2. **Discoverability**: Easy to find all code related to a specific feature.
+3. **Maintainability**: Changes to one feature are less likely to affect others.
+4. **Scalability**: New features can be added without modifying existing ones.
+5. **Testability**: Features can be tested in isolation.
+6. **Onboarding**: New developers can understand the codebase more quickly.
+
+## When to Use
+
+This pattern is particularly effective for:
+
+1. Medium to large applications with distinct business features
+2. Teams with multiple developers working on different features
+3. Applications that evolve over time with new features being added
+4. Projects that require clear boundaries between different parts of the application
+
+## Conclusion
+
+The feature-based architecture pattern provides a clean, organized approach to structuring frontend applications. By grouping code by business features rather than technical concerns, it creates a more intuitive and maintainable codebase that scales well as the application grows.
